+++ packages/utils/src/index.ts
@@ -29,7 +29,7 @@ export { Errors } from "./errors/core";
 export { StringDerived } from "./derived/string";
 export { NumberDerived } from "./derived/number";
 export { enableLogAttribute } from "./log";
-export { readingWritingSeparationDetor } from "./readingWritingSeparation/index";
+export { readingWritingSeparationDecor } from "./readingWritingSeparation/index";
 export type { readingWritingSeparationUtilsType } from "./readingWritingSeparation/index";
 export { ValidationErrorCollection, VerificationFlow } from "./verify/errorCollection";
 export { Mail, Phone, SNI, isArrayEmpty, isEmpty, isObjectEmpty } from "./verify/verify";