+++ packages/valuePolyfill/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @memo28/valuepolyfill
 
+## 1.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @memo28/types@1.1.7
+  - @memo28/utils@1.2.5
+
 ## 1.1.4
 
 ### Patch Changes